+++ include/NeoFOAM/core/executor/CPUExecutor.hpp
@@ -26,23 +26,16 @@ public:
     template<typename T>
     T* alloc(size_t size) const
     {
-        return static_cast<T*>(
-            Kokkos::kokkos_malloc<exec>("Field", size * sizeof(T))
-        );
+        return static_cast<T*>(Kokkos::kokkos_malloc<exec>("Field", size * sizeof(T)));
     }
 
     template<typename T>
     T* realloc(void* ptr, size_t newSize) const
     {
-        return static_cast<T*>(
-            Kokkos::kokkos_realloc<exec>(ptr, newSize * sizeof(T))
-        );
+        return static_cast<T*>(Kokkos::kokkos_realloc<exec>(ptr, newSize * sizeof(T)));
     }
 
-    void* alloc(size_t size) const
-    {
-        return Kokkos::kokkos_malloc<exec>("Field", size);
-    }
+    void* alloc(size_t size) const { return Kokkos::kokkos_malloc<exec>("Field", size); }
 
     void* realloc(void* ptr, size_t newSize) const
     {