+++ include/NeoFOAM/core/primitives/label.hpp
@@ -3,9 +3,15 @@
 #pragma once
 
 #include <cstdint>
+
 namespace NeoFOAM
 {
+#ifdef NEOFOAM_DP_LABEL
+using label = int64_t;
+using localIdx = uint64_t;
+#else
 using label = int32_t;
 using localIdx = uint32_t;
+#endif
 using globalIdx = uint64_t;
 }