+++ include/NeoFOAM/fields/Field.hpp
@@ -3,11 +3,14 @@
 #pragma once
 
 #include <Kokkos_Core.hpp>
-#include <iostream>
 
+#include <iostream>
 #include <span>
+
 #include "NeoFOAM/core/executor/executor.hpp"
-#include "NeoFOAM/primitives/scalar.hpp"
+#include "NeoFOAM/core/primitives/scalar.hpp"
+#include "NeoFOAM/fields/operations/OperationsMacros.hpp"
+#include "NeoFOAM/fields/FieldTypeDefs.hpp"
 
 namespace NeoFOAM
 {
@@ -29,16 +32,24 @@ public:
      * @param exec  Executor associated to the matrix
      * @param size  size of the matrix
      * */
-    Field(const Executor& exec, size_t size)
-        : size_(size), exec_(exec), data_(nullptr)
+    Field(const Executor& exec, size_t size) : size_(size), data_(nullptr), exec_(exec)
     {
         void* ptr = nullptr;
         std::visit(
-            [this, &ptr, size](const auto& exec)
-            { ptr = exec.alloc(size * sizeof(T)); },
-            exec_
+            [this, &ptr, size](const auto& exec) { ptr = exec.alloc(size * sizeof(T)); }, exec_
         );
         data_ = static_cast<T*>(ptr);
+    };
+
+    Field(const Field<T>& rhs) : size_(rhs.size_), data_(nullptr), exec_(rhs.exec_)
+    {
+        void* ptr = nullptr;
+        auto size = rhs.size_;
+        std::visit(
+            [this, &ptr, size](const auto& exec) { ptr = exec.alloc(size * sizeof(T)); }, exec_
+        );
+        data_ = static_cast<T*>(ptr);
+        setField(*this, rhs.field());
     };
 
     /**
@@ -65,10 +76,23 @@ public:
      * @brief Returns a copy of the field back to the host.
      * @returns A copy of the field on the host.
      */
-    [[nodiscard]] Field<T> copyToHost()
+    [[nodiscard]] Field<T> copyToHost() const
     {
         Field<T> result(CPUExecutor {}, size_);
-        this->copyToHost(result);
+        if (!std::holds_alternative<GPUExecutor>(exec_))
+        {
+            result = *this;
+        }
+        else
+        {
+            Kokkos::View<T*, Kokkos::DefaultExecutionSpace, Kokkos::MemoryUnmanaged> gpuView(
+                data_, size_
+            );
+            Kokkos::View<T*, Kokkos::HostSpace, Kokkos::MemoryUnmanaged> resultView(
+                result.data(), size_
+            );
+            Kokkos::deep_copy(resultView, gpuView);
+        }
         return result;
     }
 
@@ -82,22 +106,23 @@ public:
      */
     void copyToHost(Field<T>& result)
     {
+        if (result.size() != size_)
+        {
+            exit(1);
+        }
+
         if (!std::holds_alternative<GPUExecutor>(exec_))
         {
             result = *this;
         }
         else
         {
-            if (result.size() != size_)
-            {
-                exit(1);
-            }
-
-            Kokkos::
-                View<T*, Kokkos::DefaultExecutionSpace, Kokkos::MemoryUnmanaged>
-                    gpuView(data_, size_);
-            Kokkos::View<T*, Kokkos::HostSpace, Kokkos::MemoryUnmanaged>
-                resultView(result.data(), size_);
+            Kokkos::View<T*, Kokkos::DefaultExecutionSpace, Kokkos::MemoryUnmanaged> gpuView(
+                data_, size_
+            );
+            Kokkos::View<T*, Kokkos::HostSpace, Kokkos::MemoryUnmanaged> resultView(
+                result.data(), size_
+            );
             Kokkos::deep_copy(resultView, gpuView);
         }
     }
@@ -230,8 +255,7 @@ public:
     {
         void* ptr = nullptr;
         std::visit(
-            [this, &ptr, size](const auto& exec)
-            { ptr = exec.realloc(data_, size * sizeof(T)); },
+            [this, &ptr, size](const auto& exec) { ptr = exec.realloc(data_, size * sizeof(T)); },
             exec_
         );
         data_ = static_cast<T*>(ptr);
@@ -256,7 +280,7 @@ public:
      * @brief Gets the executor associated with the field.
      * @return Reference to the executor.
      */
-    [[nodiscard]] const Executor& exec() { return exec_; }
+    [[nodiscard]] const Executor& exec() const { return exec_; }
 
     /**
      * @brief Gets the size of the field.
@@ -273,21 +297,14 @@ public:
      * @brief Gets the field as a span.
      * @return Span of the field.
      */
-    [[nodiscard]] const std::span<T> field() const
-    {
-        return std::span<T>(data_, size_);
-    }
+    [[nodiscard]] const std::span<T> field() const { return std::span<T>(data_, size_); }
 
 private:
 
-    size_t size_; //!< Size of the field.
-    T* data_;     //!< Pointer to the field data.
-    const Executor
-        exec_; //!< Executor associated with the field. (CPU, GPU, openMP, etc.)
+    size_t size_;         //!< Size of the field.
+    T* data_;             //!< Pointer to the field data.
+    const Executor exec_; //!< Executor associated with the field. (CPU, GPU, openMP, etc.)
 };
 
 
 } // namespace NeoFOAM
-
-#include "FieldOperations.hpp"
-#include "FieldTypeDefs.hpp"