+++ include/NeoFOAM/helpers/exceptions.hpp
@@ -80,9 +80,8 @@ public:
         : Error(
             file,
             line,
-            func + ": Trying to perform binary operation " + " "
-                + std::to_string(lengthA) + ", " + std::to_string(lengthB) + " "
-                + clarification
+            func + ": Trying to perform binary operation " + " " + std::to_string(lengthA) + ", "
+                + std::to_string(lengthB) + " " + clarification
         )
     {}
 };
@@ -94,16 +93,11 @@ public:
  * @throw DimensionMismatch  if `_op1` and `_op2` differ in the number of
  *                           rows or columns
  */
-#define NeoFOAM_ASSERT_EQUAL_LENGTH(_op1, _op2)                                \
-    if (_op1.size() != _op2.size())                                            \
-    {                                                                          \
-        throw ::NeoFOAM::DimensionMismatch(                                    \
-            __FILE__,                                                          \
-            __LINE__,                                                          \
-            __func__,                                                          \
-            _op1.size(),                                                       \
-            _op2.size(),                                                       \
-            "expected equal dimensions"                                        \
-        );                                                                     \
+#define NeoFOAM_ASSERT_EQUAL_LENGTH(_op1, _op2)                                                    \
+    if (_op1.size() != _op2.size())                                                                \
+    {                                                                                              \
+        throw ::NeoFOAM::DimensionMismatch(                                                        \
+            __FILE__, __LINE__, __func__, _op1.size(), _op2.size(), "expected equal dimensions"    \
+        );                                                                                         \
     }
 }